ML/Feature7/requirements.txt
torch
transformers
datasets
scikit-learn
sentence-transformers
numpy
pandas
pyyaml
